+++ djangocms_blog/cms_apps.py
@@ -14,10 +14,10 @@ from .settings import get_setting
 @apphook_pool.register
 class BlogApp(AutoCMSAppMixin, CMSConfigApp):
     name = _('Blog')
-    urls = ['djangocms_blog.urls']
+    _urls = [get_setting('URLCONF')]
     app_name = 'djangocms_blog'
     app_config = BlogConfig
-    menus = [BlogCategoryMenu]
+    _menus = [BlogCategoryMenu]
     auto_setup = {
         'enabled': get_setting('AUTO_SETUP'),
         'home title': get_setting('AUTO_HOME_TITLE'),
@@ -29,6 +29,17 @@ class BlogApp(AutoCMSAppMixin, CMSConfigApp):
             'object_name': get_setting('DEFAULT_OBJECT_NAME')
         },
     }
+
+    def get_urls(self, page=None, language=None, **kwargs):
+        return [get_setting('URLCONF')]
+
+    @property
+    def urls(self):
+        return self.get_urls()
+
+    @property
+    def menus(self):
+        return self._menus
 
 
 BlogApp.setup()