+++ djangocms_blog/forms.py
@@ -3,15 +3,47 @@ from __future__ import absolute_import, print_function, unicode_literals
 
 from django import forms
 from django.conf import settings
+from django.core.validators import MaxLengthValidator
+from django.utils.functional import cached_property
 from parler.forms import TranslatableModelForm
 from taggit_autosuggest.widgets import TagAutoSuggest
+
+from djangocms_blog.settings import get_setting
 
 from .models import BlogCategory, BlogConfig, Post
 
 
-class CategoryAdminForm(TranslatableModelForm):
+class ConfigFormBase(object):
+    """
+    This provide the app_config property which returns the currently
+    selected app_config, whether it's an instance attribute or
+    passed in the request
+    """
+
+    @cached_property
+    def app_config(self):
+        if getattr(self.instance, 'app_config_id', None):
+            return self.instance.app_config
+        elif 'app_config' in self.initial:
+            return BlogConfig.objects.get(pk=self.initial['app_config'])
+        elif self.data.get('app_config', None):
+            return BlogConfig.objects.get(pk=self.data['app_config'])
+        return None
+
+
+class CategoryAdminForm(ConfigFormBase, TranslatableModelForm):
 
     def __init__(self, *args, **kwargs):
+        self.base_fields['meta_description'].validators = [
+            MaxLengthValidator(get_setting('META_DESCRIPTION_LENGTH'))
+        ]
+        original_attrs = self.base_fields['meta_description'].widget.attrs
+        if 'cols' in original_attrs:
+            del original_attrs['cols']
+        if 'rows' in original_attrs:
+            del original_attrs['rows']
+        original_attrs['maxlength'] = get_setting('META_DESCRIPTION_LENGTH')
+        self.base_fields['meta_description'].widget = forms.TextInput(original_attrs)
         super(CategoryAdminForm, self).__init__(*args, **kwargs)
 
         if 'parent' in self.fields:
@@ -20,11 +52,14 @@ class CategoryAdminForm(TranslatableModelForm):
                 qs = qs.exclude(
                     pk__in=[self.instance.pk] + [child.pk for child in self.instance.descendants()]
                 )
-
+            config = None
             if getattr(self.instance, 'app_config_id', None):
                 qs = qs.namespace(self.instance.app_config.namespace)
-            elif 'initial' in kwargs and 'app_config' in kwargs['initial']:
-                config = BlogConfig.objects.get(pk=kwargs['initial']['app_config'])
+            elif 'app_config' in self.initial:
+                config = BlogConfig.objects.get(pk=self.initial['app_config'])
+            elif self.data.get('app_config', None):
+                config = BlogConfig.objects.get(pk=self.data['app_config'])
+            if config:
                 qs = qs.namespace(config.namespace)
             self.fields['parent'].queryset = qs
 
@@ -40,32 +75,56 @@ class LatestEntriesForm(forms.ModelForm):
 
     class Media:
         css = {
-            'all': ('%sdjangocms_blog/css/%s' % (settings.STATIC_URL,
-                                                 'djangocms_blog_admin.css'),)
+            'all': ('%sdjangocms_blog/css/%s' % (
+                settings.STATIC_URL, 'djangocms_blog_admin.css'
+            ),)
         }
 
 
-class PostAdminForm(TranslatableModelForm):
+class PostAdminFormBase(ConfigFormBase, TranslatableModelForm):
+    """
+    This provide common methods between the admin and wizard form
+    """
 
     class Meta:
         model = Post
         fields = '__all__'
 
-    def __init__(self, *args, **kwargs):
-        super(PostAdminForm, self).__init__(*args, **kwargs)
-
+    @cached_property
+    def available_categories(self):
         qs = BlogCategory.objects
+        if self.app_config:
+            return qs.namespace(self.app_config.namespace).active_translations()
+        return qs
 
-        if getattr(self.instance, 'app_config_id', None):
-            qs = qs.namespace(self.instance.app_config.namespace)
-        elif 'initial' in kwargs and 'app_config' in kwargs['initial']:
-            config = BlogConfig.objects.get(pk=kwargs['initial']['app_config'])
-            qs = qs.namespace(config.namespace)
 
+class PostAdminForm(PostAdminFormBase):
+
+    def __init__(self, *args, **kwargs):
+        self.base_fields['meta_description'].validators = [
+            MaxLengthValidator(get_setting('META_DESCRIPTION_LENGTH'))
+        ]
+        original_attrs = self.base_fields['meta_description'].widget.attrs
+        if 'cols' in original_attrs:
+            del original_attrs['cols']
+        if 'rows' in original_attrs:
+            del original_attrs['rows']
+        original_attrs['maxlength'] = get_setting('META_DESCRIPTION_LENGTH')
+        self.base_fields['meta_description'].widget = forms.TextInput(original_attrs)
+        self.base_fields['meta_title'].validators = [
+            MaxLengthValidator(get_setting('META_TITLE_LENGTH'))
+        ]
+        super(PostAdminForm, self).__init__(*args, **kwargs)
         if 'categories' in self.fields:
-            self.fields['categories'].queryset = qs
+            self.fields['categories'].queryset = self.available_categories
 
         if 'app_config' in self.fields:
             # Don't allow app_configs to be added here. The correct way to add an
             # apphook-config is to create an apphook on a cms Page.
             self.fields['app_config'].widget.can_add_related = False
+
+        if self.app_config:
+            self.initial['main_image_full'] = \
+                self.app_config.app_data['config'].get('default_image_full')
+            self.initial['main_image_thumbnail'] = \
+                self.app_config.app_data['config'].get('default_image_thumbnail')